third_party/WebKit/LayoutTests/inspector/console/console-dir-deprecated.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function logObjects()
{
    console.dir(window);
    console.dir(document.getElementById("foo").createShadowRoot());
}

function test()
{
    InspectorTest.evaluateInPage("logObjects()", step2);
    function step2()
    {
        InspectorTest.dumpConsoleMessages();
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<div id="foo"></div>
<p>
Tests that console does not log deprecated warning messages while dir-dumping objects.
</p>

</body>
</html>
