third_party/WebKit/LayoutTests/inspector/console/console-dir-deprecated-expected.txt

CONSOLE MESSAGE: line 9: [object Window]
CONSOLE MESSAGE: line 10: [object ShadowRoot]
Tests that console does not log deprecated warning messages while dir-dumping objects.

Window console-dir-deprecated.html:9
#document-fragment console-dir-deprecated.html:10
